Moved environment settings to its own class
