+++ Assets/Scripts/Settings/UISettingsEnvironment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 463147a12f484cc4dbe150fb97b22392
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
